Explicitly set compute softmax in float32
The [Keras mixed precision guide](https://www.tensorflow.org/guide/keras/mixed_precision) recommends computing the sofmax in float32 for better numerical stability. If I recall correctly this was done implicitely in TensorFlow 2.0 since the softmax op was [blacklisted in the mixed precision graph rewrite](https://github.com/tensorflow/tensorflow/blob/v2.0.0/tensorflow/core/grappler/optimizers/auto_mixed_precision_lists.h#L146-L174). As of TensorFlow 2.1.0 this is no longer the case so we should follow the recommendation of the guide and explicitly set the data type.

+++ larq_zoo/literature/dorefanet.py
@@ -95,7 +95,7 @@ class DoReFaNetFactory(ModelFactory):
             out = self.fully_connected_block(out, units=4096)
             out = tf.keras.layers.Activation("clip_by_value_activation")(out)
             out = tf.keras.layers.Dense(self.num_classes, use_bias=True)(out)
-            out = tf.keras.layers.Activation("softmax")(out)
+            out = tf.keras.layers.Activation("softmax", dtype="float32")(out)
 
         model = tf.keras.Model(inputs=self.image_input, outputs=out, name="dorefanet")
